Strip leading zero's from the index
Leading zero's cause an issue during the name matching. This is
because the name matching treats each field as an atomic whole
and therefore `01` and `1` are different.

The easiest solution to the end user is to strip all the leading
zero's. This way the file maybe `01-my-guide` but the standardized
name would be `1_my_guide`.