Stop the error colourisation with a reset
All error messages are displayed as bold red. This causes issues as
the error message contains a table with its own colourisation

As a result the red/bold bleeds into the table until it hits another
ANSI colour code. The exact behaviour of this appears to be terminal
dependant.

The previous fix of colourizing everything `white` does not appear
to stop the bold seeping through. Hopefully a `reset` will work