Update to the bleeding edge of TTY tool suite
The `tty` tool suite extensively uses the a `~> 0.x` dependency
on `pastel` to provide colour support. However this has been updated
to `~> 0.8` on the master branch; where it was previously `~> 0.7`

This means all the `tty-*` libraries need to be updated to the bleeding
edge version inorder to resolve the dependencies. There was a similar
issue with the `strings` library