Add some wiggle room to TTY::Markdown text wrapping
TTY::Markdown's text wrapping can not be disabled. Instead it can be
pseudo-disabled by setting the terminal width to excessively large.

However defining "excessively large" is an open end question as
TTY::Markdown adds characters (padding, colors, etc). This is
particularly problematic for small documents which only contain a
single paragraph (and maybe a header)

A work around is to add an additional column width to the
content length. This allows TTY::Markdown to add its additional
characters without exceeding the maximum length