Support easier toggling of groff rendering
groff rendering is on by default.  It can be turned of by setting the
environment variable `flight_HOWTO_groff_render` to a "false looking"
value e.g., `false`, `off` or `0`.

This works as long as the value hasn't been set in `config.yaml`.  If it
has, the `config.yaml` value takes precedence regardless of the
environment variable setting.